WIP: Support value groups
This adds support for value groups to dig. In addition to supporting
named values with `dig.In/Out`, we now supported unordered groups of
named values.

To produce a named value, a constructor will tag a `dig.Out` field with
`group:".."`. Constructors which rely on value groups can consume these
values by declaring a slice tagged with `group:".."`.

For example, consider the YARPC TransportSpec use case. The constructor
that produces the TransportSpec will do,

    type TransportSpecResult struct {
        dig.Out

        TransportSpec config.TransportSpec `group:"yarpc.transports"`
    }

    func NewTransportSpec() TransportSpecResult

The constructor that consumes the TransportSpecs will do,

    type Param struct {
        dig.In

        TransportSpecs []config.TransportSpec `group:"yarpc.transports"`
    }

    func NewThing(Param) *Thing