node: Track if already called
Now that nodes represent individual constructors in the graph rather
than the types they provide, we can track on the node whether it was
already called.

This has no value right now but we will need this once we introduce
value groups to ensure that constructors don't get called multiple
times.