result: Find results with resultVisitor
Similar to paramVisitor, this adds resultVisitor and walkResult. In
addition to visiting results, resultVisitor also knows which result
position or result object field it is visiting, allowing for more
detailed error messages.

For example, this lets us provide a lot more information in the case
when the constructor is providing the same type multiple times. Given,

    type out struct {
        dig.Out

        Foo struct {
            dig.Out

            Writer io.Writer
        }

        Bar struct {
            dig.Out

            Writer io.Writer
        }
    }

Instead of,

   returns multiple io.Writer

We will now get,

    cannot provide io.Writer from [0].Bar.Writer in constructor func() out:
        already provided by [0].Foo.Writer