graph: Support multiple producers for a type
This changes how the graph is represented internally from storing a
`map[key]*node` to `map[key][]*node`. This essentially allows a key to
have a dependency on multiple nodes (constructors).

As part of this change, I also renamed this dependency map to
`producers` and the store for values from `cache` to `values`, which I
think is clearer.